+++ 20170919/刘驭壬-141242026/src/hulu/Main.java
@@ -1,16 +1,17 @@
+package hulu;
+
 public class Main {
 
     public static void main(String[] args) {
 	// write your code here
         Hulubros h = new Hulubros();
         h.printName();
-        Sort s = new Sort();
-        s.bubbleSort(h);
+        h.bubbleSort();
         h.printName();
         h.printColor();
         h.shuffle();
         h.printName();
-        s.quickSort(h);
+        h.quickSort();
         h.printName();
     }
 }
